Split constraint tests into its own file
